move lib bridge to env
