+++ rocksdb-cxx/db/error_handler.h
@@ -11,9 +11,9 @@
 #include "rocksdb/status.h"
 
 #ifndef ROCKSDB_RS
-#include "rocksdb-rs-cxx/lib.h"
+#include "rocksdb-rs-cxx/env.h"
 #else
-#include "rocksdb-rs/src/lib.rs.h"
+#include "rocksdb-rs/src/env.rs.h"
 #endif
 
 namespace ROCKSDB_NAMESPACE {
