+++ rocksdb-cxx/file/sst_file_manager_impl.h
@@ -14,9 +14,9 @@
 #include "rocksdb/sst_file_manager.h"
 
 #ifndef ROCKSDB_RS
-#include "rocksdb-rs-cxx/lib.h"
+#include "rocksdb-rs-cxx/env.h"
 #else
-#include "rocksdb-rs/src/lib.rs.h"
+#include "rocksdb-rs/src/env.rs.h"
 #endif
 
 namespace ROCKSDB_NAMESPACE {
