+++ rocksdb-cxx/util/status.cc
@@ -18,9 +18,9 @@
 #include "port/port.h"
 
 #ifndef ROCKSDB_RS
-#include "rocksdb-rs-cxx/lib.h"
+#include "rocksdb-rs-cxx/env.h"
 #else
-#include "rocksdb-rs/src/lib.rs.h"
+#include "rocksdb-rs/src/env.rs.h"
 #endif
 
 namespace ROCKSDB_NAMESPACE {
